fix: Properly handle `fromBytes` for PrngTransaction
