+++ Sources/Hedera/AnyTransaction.swift
@@ -20,366 +20,6 @@
 
 import HederaProtobufs
 
-internal enum AnyTransaction {
-    case accountCreate(AccountCreateTransaction)
-    case accountUpdate(AccountUpdateTransaction)
-    case accountDelete(AccountDeleteTransaction)
-    case accountAllowanceApprove(AccountAllowanceApproveTransaction)
-    case accountAllowanceDelete(AccountAllowanceDeleteTransaction)
-    case contractCreate(ContractCreateTransaction)
-    case contractUpdate(ContractUpdateTransaction)
-    case contractDelete(ContractDeleteTransaction)
-    case contractExecute(ContractExecuteTransaction)
-    case transfer(TransferTransaction)
-    case topicCreate(TopicCreateTransaction)
-    case topicUpdate(TopicUpdateTransaction)
-    case topicDelete(TopicDeleteTransaction)
-    case topicMessageSubmit(TopicMessageSubmitTransaction)
-    case fileAppend(FileAppendTransaction)
-    case fileCreate(FileCreateTransaction)
-    case fileUpdate(FileUpdateTransaction)
-    case fileDelete(FileDeleteTransaction)
-    case prng(PrngTransaction)
-    case tokenAssociate(TokenAssociateTransaction)
-    case tokenBurn(TokenBurnTransaction)
-    case tokenCreate(TokenCreateTransaction)
-    case tokenDelete(TokenDeleteTransaction)
-    case tokenDissociate(TokenDissociateTransaction)
-    case tokenFeeScheduleUpdate(TokenFeeScheduleUpdateTransaction)
-    case tokenFreeze(TokenFreezeTransaction)
-    case tokenGrantKyc(TokenGrantKycTransaction)
-    case tokenMint(TokenMintTransaction)
-    case tokenPause(TokenPauseTransaction)
-    case tokenRevokeKyc(TokenRevokeKycTransaction)
-    case tokenUnfreeze(TokenUnfreezeTransaction)
-    case tokenUnpause(TokenUnpauseTransaction)
-    case tokenUpdate(TokenUpdateTransaction)
-    case tokenWipe(TokenWipeTransaction)
-    case systemDelete(SystemDeleteTransaction)
-    case systemUndelete(SystemUndeleteTransaction)
-    case freeze(FreezeTransaction)
-    case scheduleCreate(ScheduleCreateTransaction)
-    case scheduleSign(ScheduleSignTransaction)
-    case scheduleDelete(ScheduleDeleteTransaction)
-    case ethereum(EthereumTransaction)
-
-    internal init(upcasting transaction: Transaction) {
-        switch transaction {
-        case let transaction as AccountCreateTransaction: self = .accountCreate(transaction)
-        case let transaction as AccountUpdateTransaction: self = .accountUpdate(transaction)
-        case let transaction as AccountDeleteTransaction: self = .accountDelete(transaction)
-        case let transaction as AccountAllowanceApproveTransaction: self = .accountAllowanceApprove(transaction)
-        case let transaction as AccountAllowanceDeleteTransaction: self = .accountAllowanceDelete(transaction)
-        case let transaction as ContractCreateTransaction: self = .contractCreate(transaction)
-        case let transaction as ContractUpdateTransaction: self = .contractUpdate(transaction)
-        case let transaction as ContractDeleteTransaction: self = .contractDelete(transaction)
-        case let transaction as ContractExecuteTransaction: self = .contractExecute(transaction)
-        case let transaction as TransferTransaction: self = .transfer(transaction)
-        case let transaction as TopicCreateTransaction: self = .topicCreate(transaction)
-        case let transaction as TopicUpdateTransaction: self = .topicUpdate(transaction)
-        case let transaction as TopicDeleteTransaction: self = .topicDelete(transaction)
-        case let transaction as TopicMessageSubmitTransaction: self = .topicMessageSubmit(transaction)
-        case let transaction as FileAppendTransaction: self = .fileAppend(transaction)
-        case let transaction as FileCreateTransaction: self = .fileCreate(transaction)
-        case let transaction as FileUpdateTransaction: self = .fileUpdate(transaction)
-        case let transaction as FileDeleteTransaction: self = .fileDelete(transaction)
-        case let transaction as TokenAssociateTransaction: self = .tokenAssociate(transaction)
-        case let transaction as TokenBurnTransaction: self = .tokenBurn(transaction)
-        case let transaction as TokenCreateTransaction: self = .tokenCreate(transaction)
-        case let transaction as TokenDeleteTransaction: self = .tokenDelete(transaction)
-        case let transaction as TokenDissociateTransaction: self = .tokenDissociate(transaction)
-        case let transaction as TokenFeeScheduleUpdateTransaction: self = .tokenFeeScheduleUpdate(transaction)
-        case let transaction as TokenFreezeTransaction: self = .tokenFreeze(transaction)
-        case let transaction as TokenGrantKycTransaction: self = .tokenGrantKyc(transaction)
-        case let transaction as TokenMintTransaction: self = .tokenMint(transaction)
-        case let transaction as TokenPauseTransaction: self = .tokenPause(transaction)
-        case let transaction as TokenRevokeKycTransaction: self = .tokenRevokeKyc(transaction)
-        case let transaction as TokenUnfreezeTransaction: self = .tokenUnfreeze(transaction)
-        case let transaction as TokenUnpauseTransaction: self = .tokenUnpause(transaction)
-        case let transaction as TokenUpdateTransaction: self = .tokenUpdate(transaction)
-        case let transaction as TokenWipeTransaction: self = .tokenWipe(transaction)
-        case let transaction as SystemDeleteTransaction: self = .systemDelete(transaction)
-        case let transaction as SystemUndeleteTransaction: self = .systemUndelete(transaction)
-        case let transaction as FreezeTransaction: self = .freeze(transaction)
-        case let transaction as ScheduleCreateTransaction: self = .scheduleCreate(transaction)
-        case let transaction as ScheduleSignTransaction: self = .scheduleSign(transaction)
-        case let transaction as ScheduleDeleteTransaction: self = .scheduleDelete(transaction)
-        case let transaction as EthereumTransaction: self = .ethereum(transaction)
-        default: fatalError("Unrecognized transaction type")
-        }
-    }
-
-    // internal enum Kind: String {
-    //     case accountCreate
-    //     case accountUpdate
-    //     case accountDelete
-    //     case accountAllowanceApprove
-    //     case accountAllowanceDelete
-    //     case contractCreate
-    //     case contractUpdate
-    //     case contractDelete
-    //     case contractExecute
-    //     case transfer
-    //     case topicCreate
-    //     case topicUpdate
-    //     case topicDelete
-    //     case topicMessageSubmit
-    //     case fileAppend
-    //     case fileCreate
-    //     case fileUpdate
-    //     case fileDelete
-    //     case tokenAssociate
-    //     case tokenBurn
-    //     case tokenCreate
-    //     case tokenDelete
-    //     case tokenDissociate
-    //     case tokenFeeScheduleUpdate
-    //     case tokenFreeze
-    //     case tokenGrantKyc
-    //     case tokenMint
-    //     case tokenPause
-    //     case tokenRevokeKyc
-    //     case tokenUnfreeze
-    //     case tokenUnpause
-    //     case tokenUpdate
-    //     case tokenWipe
-    //     case systemDelete
-    //     case systemUndelete
-    //     case freeze
-    //     case scheduleCreate
-    //     case scheduleSign
-    //     case scheduleDelete
-    //     case ethereum
-    // }
-
-    internal var transaction: Transaction {
-        switch self {
-        case .accountCreate(let transaction): return transaction
-        case .accountUpdate(let transaction): return transaction
-        case .accountDelete(let transaction): return transaction
-        case .accountAllowanceApprove(let transaction): return transaction
-        case .accountAllowanceDelete(let transaction): return transaction
-        case .contractCreate(let transaction): return transaction
-        case .contractUpdate(let transaction): return transaction
-        case .contractDelete(let transaction): return transaction
-        case .contractExecute(let transaction): return transaction
-        case .transfer(let transaction): return transaction
-        case .topicCreate(let transaction): return transaction
-        case .topicUpdate(let transaction): return transaction
-        case .topicDelete(let transaction): return transaction
-        case .topicMessageSubmit(let transaction): return transaction
-        case .fileAppend(let transaction): return transaction
-        case .fileCreate(let transaction): return transaction
-        case .fileUpdate(let transaction): return transaction
-        case .fileDelete(let transaction): return transaction
-        case .tokenAssociate(let transaction): return transaction
-        case .tokenBurn(let transaction): return transaction
-        case .tokenCreate(let transaction): return transaction
-        case .tokenDelete(let transaction): return transaction
-        case .tokenDissociate(let transaction): return transaction
-        case .tokenFeeScheduleUpdate(let transaction): return transaction
-        case .tokenFreeze(let transaction): return transaction
-        case .tokenGrantKyc(let transaction): return transaction
-        case .tokenMint(let transaction): return transaction
-        case .tokenPause(let transaction): return transaction
-        case .tokenRevokeKyc(let transaction): return transaction
-        case .tokenUnfreeze(let transaction): return transaction
-        case .tokenUnpause(let transaction): return transaction
-        case .tokenUpdate(let transaction): return transaction
-        case .tokenWipe(let transaction): return transaction
-        case .systemDelete(let transaction): return transaction
-        case .systemUndelete(let transaction): return transaction
-        case .freeze(let transaction): return transaction
-        case .scheduleCreate(let transaction): return transaction
-        case .scheduleSign(let transaction): return transaction
-        case .scheduleDelete(let transaction): return transaction
-        case .ethereum(let transaction): return transaction
-        case .prng(let transaction): return transaction
-        }
-    }
-}
-
-extension AnyTransaction {
-    internal static func fromProtobuf(
-        _ firstBody: Proto_TransactionBody, _ data: [Proto_TransactionBody.OneOf_Data]
-    ) throws -> Self {
-        func intoOnlyValue<Element>(_ array: [Element]) throws -> Element {
-            guard array.count == 1 else {
-                throw HError.fromProtobuf("chunks in non chunkable transaction")
-            }
-
-            return array[0]
-        }
-
-        let data = try ServicesTransactionDataList.fromProtobuf(data)
-
-        switch data {
-
-        case .accountCreate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try AccountCreateTransaction(protobuf: firstBody, value))
-
-        case .accountUpdate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try AccountUpdateTransaction(protobuf: firstBody, value))
-
-        case .accountDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try AccountDeleteTransaction(protobuf: firstBody, value))
-
-        case .accountAllowanceApprove(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try AccountAllowanceApproveTransaction(protobuf: firstBody, value))
-
-        case .accountAllowanceDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try AccountAllowanceDeleteTransaction(protobuf: firstBody, value))
-
-        case .contractCreate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ContractCreateTransaction(protobuf: firstBody, value))
-
-        case .contractUpdate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ContractUpdateTransaction(protobuf: firstBody, value))
-
-        case .contractDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ContractDeleteTransaction(protobuf: firstBody, value))
-
-        case .contractExecute(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ContractExecuteTransaction(protobuf: firstBody, value))
-
-        case .transfer(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TransferTransaction(protobuf: firstBody, value))
-
-        case .topicCreate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TopicCreateTransaction(protobuf: firstBody, value))
-
-        case .topicUpdate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TopicUpdateTransaction(protobuf: firstBody, value))
-
-        case .topicDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TopicDeleteTransaction(protobuf: firstBody, value))
-
-        case .topicMessageSubmit(let value):
-            return Self(upcasting: try TopicMessageSubmitTransaction(protobuf: firstBody, value))
-
-        case .fileAppend(let value):
-            return Self(upcasting: try FileAppendTransaction(protobuf: firstBody, value))
-
-        case .fileCreate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try FileCreateTransaction(protobuf: firstBody, value))
-
-        case .fileUpdate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try FileUpdateTransaction(protobuf: firstBody, value))
-
-        case .fileDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try FileDeleteTransaction(protobuf: firstBody, value))
-
-        case .tokenAssociate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenAssociateTransaction(protobuf: firstBody, value))
-
-        case .tokenBurn(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenBurnTransaction(protobuf: firstBody, value))
-
-        case .tokenCreate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenCreateTransaction(protobuf: firstBody, value))
-
-        case .tokenDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenDeleteTransaction(protobuf: firstBody, value))
-
-        case .tokenDissociate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenDissociateTransaction(protobuf: firstBody, value))
-
-        case .tokenFeeScheduleUpdate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenFeeScheduleUpdateTransaction(protobuf: firstBody, value))
-
-        case .tokenFreeze(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenFreezeTransaction(protobuf: firstBody, value))
-
-        case .tokenGrantKyc(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenGrantKycTransaction(protobuf: firstBody, value))
-
-        case .tokenMint(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenMintTransaction(protobuf: firstBody, value))
-
-        case .tokenPause(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenPauseTransaction(protobuf: firstBody, value))
-
-        case .tokenRevokeKyc(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenRevokeKycTransaction(protobuf: firstBody, value))
-
-        case .tokenUnfreeze(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenUnfreezeTransaction(protobuf: firstBody, value))
-
-        case .tokenUnpause(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenUnpauseTransaction(protobuf: firstBody, value))
-
-        case .tokenUpdate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenUpdateTransaction(protobuf: firstBody, value))
-
-        case .tokenWipe(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try TokenWipeTransaction(protobuf: firstBody, value))
-
-        case .systemDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try SystemDeleteTransaction(protobuf: firstBody, value))
-
-        case .systemUndelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try SystemUndeleteTransaction(protobuf: firstBody, value))
-
-        case .freeze(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try FreezeTransaction(protobuf: firstBody, value))
-
-        case .scheduleCreate(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ScheduleCreateTransaction(protobuf: firstBody, value))
-
-        case .scheduleSign(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ScheduleSignTransaction(protobuf: firstBody, value))
-
-        case .scheduleDelete(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try ScheduleDeleteTransaction(protobuf: firstBody, value))
-
-        case .ethereum(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try EthereumTransaction(protobuf: firstBody, value))
-
-        case .prng(let value):
-            let value = try intoOnlyValue(value)
-            return Self(upcasting: try PrngTransaction(protobuf: firstBody, value))
-        }
-    }
-}
-
 // exists for the same reason as rust and still sucks :/\
 internal enum ServicesTransactionDataList {
     case accountCreate([Proto_CryptoCreateTransactionBody])
